pyproject.toml
[project]
name = "challeng1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.35.0",
]
